requirements.txt
flask
gunicorn
numpy
joblib
scikit-learn
google-genai